requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cssutils==2.7.1
cycler==0.12.0
dataframe-image==0.2.1
defusedxml==0.7.1
exceptiongroup==1.1.3
fastjsonschema==2.18.0
fonttools==4.43.0
frozenlist==1.4.0
h11==0.14.0
html2image==2.0.4.3
httpcore==0.17.3
httpx==0.24.1
idna==3.4
imgkit==1.2.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.5
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.0
mistune==3.0.1
multidict==6.0.4
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
numpy==1.26.0
packaging==23.1
pandas==1.5.3
pandocfilters==1.5.0
Pillow==10.0.1
platformdirs==3.10.0
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-telegram-bot==20.5
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.10.0
urllib3==2.0.5
webencodings==0.5.1
websocket-client==1.6.3
yarl==1.9.2
zipp==3.17.0
